requirements.txt
attrs==18.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.7.0.1
backports.weakref==1.0rc1
bleach==1.5.0
bz2file==0.98
certifi==2019.6.16
cffi==1.12.3
cftime==1.0.2.1
chardet==3.0.4
cloudpickle==1.2.1
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.0
defusedxml==0.6.0
docker==4.4.4
easydict==1.9
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
h5py==2.10.0
html5lib==0.9999999
hypothesis==3.82.1
idna==2.8
importlib-metadata==1.5.0
imutils==0.5.3
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.3
kiwisolver==1.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.4
mkl-service==2.0.2
mock==3.0.5
mxnet-cu80==1.3.1
nbconvert==5.6.1
nbformat==4.4.0
netCDF4==1.4.2
networkx==2.2
nibabel==2.5.0
niftidataset==0.1.5
notebook==5.7.8
numpy==1.15.4
nvidia-ml-py==375.53.1
open3d-python==0.7.0.0
opencv-python==4.1.0.25
panda==0.3.1
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pip==19.2.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pycparser==2.19
pydensecrf==1.0rc3
pydicom==1.4.2
Pygments==2.5.2
pynvx==1.0.0
pyparsing==2.2.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2018.4
PyWavelets==1.0.3
PyYAML==3.12
pyzmq==19.0.0
requests==2.21.0
scandir==1.10.0
scikit-image==0.14.3
scikit-learn==0.20.4
scipy==1.2.2
Send2Trash==1.5.0
setuptools==39.2.0
simplegeneric==0.8.1
SimpleITK==1.2.4
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
SPARNN==0.1.dev0
subprocess32==3.5.4
synthtorch==0.3.9
tensorflow-gpu==1.2.0
tensorflow-tensorboard==0.4.0
terminado==0.8.3
testpath==0.4.4
torch==1.0.0
torchsummary==1.5.1
torchvision==0.4.0
tornado==5.1.1
traitlets==4.3.3
typing==3.6.6
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.58.0
Werkzeug==1.0.0
wget==3.2
wheel==0.33.4
widgetsnbextension==3.5.1
zipp==1.2.0
